test/test_renderer.py
import unittest
from pyzplrender.renderer import ZPLRenderer
import os
from PIL import Image

class TestZPLRenderer(unittest.TestCase):
    def setUp(self):
        # 100mm x 100mm label for testing (800x800 dots)
        self.renderer = ZPLRenderer(width_mm=100, height_mm=100, dpmm=8)

    def tearDown(self):
        # Clean up any generated test images
        for f in ["test_basic.png", "test_shapes.png", "test_barcode.png", 
                  "test_qrcode.png", "test_gf.png", "test_fonts.py"]:
            if os.path.exists(f):
                try:
                    os.remove(f)
                except:
                    pass

    def test_basic_text(self):
        zpl = "^XA^FO50,50^FDTest Text^FS^XZ"
        self.renderer.render(zpl)
        self.renderer.save_png("test_basic.png")
        self.assertTrue(os.path.exists("test_basic.png"))

    def test_shapes_and_reverse(self):
        # Test Box, Circle, and Field Reverse
        zpl = (
            "^XA"
            "^FO50,100^GB300,100,4^FS"
            "^FO400,100^GC100,4^FS"
            "^FO50,250^FR^GB100,50,2^FS"  # Reversed box
            "^FO50,350^FR^FDReversed Text^FS"
            "^XZ"
        )
        self.renderer.render(zpl)
        self.renderer.save_png("test_shapes.png")
        self.assertTrue(os.path.exists("test_shapes.png"))

    def test_barcode_128(self):
        # Test Code 128 with data prefixes
        zpl = "^XA^FO50,100^BCN,100,Y,N,N^FD>:123456^FS^XZ"
        self.renderer.render(zpl)
        self.renderer.save_png("test_barcode.png")
        self.assertTrue(os.path.exists("test_barcode.png"))

    def test_qrcode(self):
        # Test QR Code with typical prefix
        zpl = "^XA^FO50,100^BQN,2,6^FDQA,https://example.com^FS^XZ"
        self.renderer.render(zpl)
        self.renderer.save_png("test_qrcode.png")
        self.assertTrue(os.path.exists("test_qrcode.png"))

    def test_graphic_field(self):
        # Test 8x8 black square in hex
        zpl = "^XA^FO50,50^GFA,8,8,1,FFFFFFFFFFFFFFFF^FS^XZ"
        self.renderer.render(zpl)
        self.renderer.save_png("test_gf.png")
        self.assertTrue(os.path.exists("test_gf.png"))

    def test_graphic_field_invalid(self):
        # Test malformed hex data shouldn't crash
        zpl = "^XA^FO50,50^GFA,8,8,1,ZZZZ^FS^XZ"
        try:
            self.renderer.render(zpl)
        except Exception as e:
            self.fail(f"Renderer crashed on invalid hex: {e}")

    def test_font_scaling(self):
        # Test font scaling command
        zpl = (
            "^XA"
            "^FO50,50^A0,N,80,80^FDLarge^FS"
            "^FO50,150^A0,N,20,20^FDSmall^FS"
            "^XZ"
        )
        self.renderer.render(zpl)
        self.assertEqual(self.renderer.active_font_height, 20)
        self.assertEqual(self.renderer.active_font_width, 20)

    def test_label_home(self):
        # ^LH100,100 should make ^FO50,50 appear at 150,150
        zpl = "^XA^LH100,100^FO50,50^FDHomeOffset^FS^XZ"
        self.renderer.render(zpl)
        self.assertEqual(self.renderer.current_x, 150)
        self.assertEqual(self.renderer.current_y, 150)

    def test_change_default_font(self):
        # ^CF should change active font height
        zpl = "^XA^CF0,60,60^FDDefaultFont^FS^XZ"
        self.renderer.render(zpl)
        self.assertEqual(self.renderer.active_font_height, 60)

    def test_rotation(self):
        # ^A0R should set orientation to Rotated (90 deg)
        zpl = "^XA^FO100,100^A0R,40,40^FDRotatedText^FS^XZ"
        self.renderer.render(zpl)
        self.assertEqual(self.renderer.current_orientation, 'R')

    def test_field_block(self):
        # Test wrapping and centering
        zpl = "^XA^FO50,50^FB300,3,0,C^FDThis is a long text that should wrap and be centered^FS^XZ"
        self.renderer.render(zpl)
        self.assertEqual(self.renderer.block_justification, 'L') # Reset after render? 
        # Actually, FO resets it.
        self.renderer.save_png("test_block.png")
        self.assertTrue(os.path.exists("test_block.png"))
        os.remove("test_block.png")

    def test_rounded_box(self):
        # ^GB300,100,4,,3 should use rounded corners
        zpl = "^XA^FO50,100^GB300,100,4,,3^FS^XZ"
        self.renderer.render(zpl)
        # Verify it doesn't crash and generates output
        self.renderer.save_png("test_rounded.png")
        self.assertTrue(os.path.exists("test_rounded.png"))
        os.remove("test_rounded.png")

    def test_context_lookback(self):
        # Ensure _get_last_meaningful_cmd works across gaps
        zpl = "^XA^BC^FO100,100^FDData^FS^XZ"
        self.renderer.render(zpl)
        # The internal current_cmd_idx will be at ^FD (index 3)
        # _get_last_meaningful_cmd should find ^BC (index 1) skipping ^FO (index 2)
        # We can't easily test private methods directly without index manipulation 
        # but the fact that the barcode renders correctly implies it works.

if __name__ == "__main__":
    unittest.main()
