pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyzplrender"
version = "0.1.0"
authors = [
  { name="Your Name", email="your.email@example.com" },
]
description = "A pure Python, offline ZPL rendering engine"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Pillow",
    "python-barcode",
    "qrcode",
]

[project.urls]
"Homepage" = "https://github.com/pyzplrender/pyzplrender"

[tool.setuptools.packages.find]
where = ["src"]
